refactor(client): add @/types folder for client-only shared types
Relocate the pure, runtime-free shared type modules (ControlledDialogProps,
EditRowBaseProps, BaseFieldProps) out of components/ into a dedicated
src/types/ folder, giving client-only types a documented home distinct from
the cross-package @emstack/types source of truth. Update the client CLAUDE.md
to spell out the distinction and which types belong where.
