+++ packages/client/src/components/formFields/NumberField.tsx
@@ -1,4 +1,4 @@
-import type { BaseFieldProps } from "@/components/formFields/fieldProps";
+import type { BaseFieldProps } from "@/types/fieldProps";
 
 import { Field, FieldError, FieldLabel } from "@/components/forms/field";
 import { Input } from "@/components/input";
