+++ packages/client/src/components/quickAdd/QuickAddRoutineDialog.tsx
@@ -1,4 +1,4 @@
-import type { ControlledDialogProps } from "@/components/dialogProps";
+import type { ControlledDialogProps } from "@/types/dialogProps";
 import type { RoutineMode } from "@emstack/types";
 
 import { QuickAddDialogFooter } from "./QuickAddDialogFooter";
