+++ packages/client/src/components/quickAdd/QuickAddTodoistDialog.tsx
@@ -1,4 +1,4 @@
-import type { ControlledDialogProps } from "@/components/dialogProps";
+import type { ControlledDialogProps } from "@/types/dialogProps";
 
 import { QuickAddIntegrationDialog } from "./QuickAddIntegrationDialog";
 import { useQuickAddTodoist } from "./useQuickAddTodoist";
