+++ packages/client/src/components/quickAdd/useQuickAddReadwise.ts
@@ -1,4 +1,4 @@
-import type { ControlledDialogProps } from "@/components/dialogProps";
+import type { ControlledDialogProps } from "@/types/dialogProps";
 
 import { useEffect, useState } from "react";
