+++ packages/client/src/components/quickAdd/useQuickAddRoutine.ts
@@ -1,4 +1,4 @@
-import type { ControlledDialogProps } from "@/components/dialogProps";
+import type { ControlledDialogProps } from "@/types/dialogProps";
 import type { RoutineMode } from "@emstack/types";
 
 import { useEffect, useState } from "react";
