+++ packages/client/src/routes/dashboard.-components/-AddLayoutDialog.tsx
@@ -1,4 +1,4 @@
-import type { ControlledDialogProps } from "@/components/dialogProps";
+import type { ControlledDialogProps } from "@/types/dialogProps";
 import type { DashboardLayout, DashboardLayoutTile } from "@emstack/types";
 
 import { useEffect, useMemo, useState } from "react";
