+++ packages/client/src/routes/dashboard.-components/-VisibleTilesDialog.tsx
@@ -1,4 +1,4 @@
-import type { ControlledDialogProps } from "@/components/dialogProps";
+import type { ControlledDialogProps } from "@/types/dialogProps";
 import type { DashboardLayout, DashboardTileId } from "@emstack/types";
 
 import { Checkbox } from "@/components/ui/checkbox";
